Address NickGeneva review comments
- CHANGELOG: remove implied GHCNHourlyLexicon entry; move GHCNLexicon rename to Changed
- ghcn.py: make get_station_metadata call sync (drop asyncio.to_thread)
- ghcn.py: cast float schema columns to float32 in _create_observation_dataframe
- ghcn.py: deduplicate observations by (STATION, DATE) in GHCNHourly fetch
- ghcn.py: deduplicate observations by (ID, DATE) per variable in GHCNDaily fetch
- test_ghcn_daily.py: move TestGHCNDailyLexicon to test/lexicon/test_ghcn_lexicon.py
- test_ghcn_lexicon.py: add TestGHCNDailyLexicon (moved from test_ghcn_daily.py)
