+++ CHANGELOG.md
@@ -12,12 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added GHCN hourly data source (`GHCNHourly`), superseding the deprecated ISD source
-- Added `GHCNHourlyLexicon` for GHCN hourly. Renamed `GHCNLexicon` to `GHCNDailyLexicon`
 - Added EarthMover ERA5 0.25 degree reanalysis data source
 - Added EarthMover IFS 0.1 degree data source and forecast source hosted by BrightBand
 
 ### Changed
 
+- Renamed `GHCNLexicon` to `GHCNDailyLexicon` for consistency with the new hourly lexicon
 - Updated MeteosatFCI reader and lexicon to include all channels.
 - Updated StormScope model package to use improved higher resolution checkpoints. Model
   now defaults to using 3 km and 10 minute spatiotemporal resolution, and includes
