+++ test/data/test_ghcn_daily.py
@@ -26,7 +26,6 @@ import pyarrow.parquet as pq
 import pytest
 
 from earth2studio.data import GHCNDaily
-from earth2studio.lexicon.ghcn import GHCNDailyLexicon
 
 
 @pytest.mark.slow
@@ -275,59 +274,6 @@ def test_ghcn_available():
 
     # Future time — partition unlikely to exist
     assert not GHCNDaily.available(datetime(2099, 1, 1))
-
-
-class TestGHCNDailyLexicon:
-    @pytest.mark.parametrize(
-        "var, element",
-        [
-            ("t2m_max", "TMAX"),
-            ("t2m_min", "TMIN"),
-            ("t2m", "TAVG"),
-            ("d2m", "ADPT"),
-            ("r2m", "RHAV"),
-            ("tp", "PRCP"),
-            ("sf", "WESF"),
-            ("sd", "SNWD"),
-            ("sde", "SNOW"),
-            ("ws10m", "AWND"),
-            ("fg10m", "WSF2"),
-            ("tcc", "ACMH"),
-        ],
-    )
-    def test_element_map(self, var, element):
-        assert GHCNDailyLexicon.VOCAB[var] == element
-
-    def test_lexicon_keys(self):
-        for var in GHCNDailyLexicon.VOCAB:
-            element, mod = GHCNDailyLexicon[var]
-            assert element == GHCNDailyLexicon.VOCAB[var]
-            assert callable(mod)
-
-    @pytest.mark.parametrize(
-        "var, raw, expected",
-        [
-            ("t2m_max", np.array([200.0]), np.array([293.15])),  # 20.0 C -> K
-            ("t2m_min", np.array([-100.0]), np.array([263.15])),  # -10.0 C -> K
-            ("t2m", np.array([200.0]), np.array([293.15])),  # 20.0 C -> K
-            ("d2m", np.array([150.0]), np.array([288.15])),  # 15.0 C -> K
-            ("r2m", np.array([75.0]), np.array([75.0])),  # percent (no conv)
-            ("tp", np.array([100.0]), np.array([0.01])),  # 10.0 mm -> 0.01 m
-            ("sf", np.array([100.0]), np.array([0.01])),  # 10.0 mm -> 0.01 m
-            ("sd", np.array([500.0]), np.array([0.5])),  # 500 mm -> 0.5 m
-            ("sde", np.array([250.0]), np.array([0.25])),  # 250 mm -> 0.25 m
-            ("ws10m", np.array([50.0]), np.array([5.0])),  # 50 tenths -> 5 m/s
-            ("fg10m", np.array([120.0]), np.array([12.0])),  # 120 tenths -> 12 m/s
-            ("tcc", np.array([80.0]), np.array([0.8])),  # 80% -> 0.8 fraction
-        ],
-    )
-    def test_unit_conversions(self, var, raw, expected):
-        _, mod = GHCNDailyLexicon[var]
-        np.testing.assert_allclose(mod(raw), expected, atol=1e-6)
-
-    def test_invalid_variable(self):
-        with pytest.raises(KeyError):
-            GHCNDailyLexicon["nonexistent"]
 
 
 class TestGHCNMock:
